Removes dead code from diagnostics presenters
