--- lib/ExecutionControlEpic/DiagnosticsFeature/Actions/OpenDiagnosticsTraveller.js
@@ -1,18 +1,0 @@
-"use babel";
-// @flow
-
-export type OpenDiagnosticsTravellerAction = () => void;
-import { renderDiagnosticsPanel } from "../AtomLinks/Panels";
-import type { PlanConfig } from "../../PlanConfigurationFeature/Types/types";
-import type { DiagnosticMode } from "../Types/types";
-
-export function openDiagnosticsTraveller(
-  toolId: string,
-  toolName: string,
-  defaultDiagnosticsMode: DiagnosticMode,
-  plan: PlanConfig,
-): OpenDiagnosticsTravellerAction {
-  return () => {
-    renderDiagnosticsPanel(toolId, toolName, defaultDiagnosticsMode, plan);
-  };
-}
