+++ lib/ExecutionControlEpic/DiagnosticsFeature/Presenters/DiagnosticsPanel.js
@@ -4,36 +4,11 @@
 import * as React from "react";
 import { compose, withProps, withState } from "recompose";
 import styled from "styled-components";
-import DiagnosticsModeSelector from "./DiagnosticsModeSelector";
 import { List, Map, Set } from "immutable";
 import DiagnosticDetailsFlow from "./DiagnosticDetailsFlow";
 import type { DiagnosticSeverity, MoleculeDiagnostic } from "../Types/types";
 import type { Range } from "../../LanguageServerProtocolFeature/Types/standard";
-import Term from "../../ActionSystemFeature/Presenters/Term";
 import DiagnosticsQuickSelector from "./DiagnosticsQuickSelector";
-import DiagnosticsFilters from "./DiagnosticsFilters";
-import type { Task } from "../../TaskExecutionFeature/Types/types";
-
-const Panel = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: stretch;
-  flex: 1;
-`;
-
-const ContentBox = styled.div`
-  display: flex;
-  align-items: stretch;
-  flex: 1;
-  flex-direction: column;
-  padding: 8px 0px 8px 8px;
-`;
-
-const ModeBox = styled.div`
-  display: flex;
-  align-items: stretch;
-  flex-direction: column;
-`;
 
 export type TravellingState = {
   diagnosticsTypes: Set<DiagnosticSeverity>,
@@ -225,116 +200,3 @@ export const DiagnosticsContent = compose(
   Filter,
   ContentModeAdapter,
 )(DiagnosticsContentPanel);
-
-function DiagnosticsPanel({
-  diagnostics = Map(),
-  mode,
-  onModeSelection,
-  onJumpTo,
-  xtermInstance,
-  currentTask,
-}: {
-  diagnostics: Map<string, Map<DiagnosticSeverity, List<MoleculeDiagnostic>>>,
-  mode: PanelMode,
-  onModeSelection: (mode: PanelMode) => void,
-  onJumpTo: (path: string, range: Range, pending: boolean) => void,
-  xtermInstance: any,
-  tasks: List<Task>,
-  onTaskClick: (task: Task) => void,
-  currentTask: Task,
-}) {
-  return (
-    <Panel>
-      <ModeBox>
-        <DiagnosticsModeSelector
-          type={mode}
-          onClick={onModeSelection}
-          terminal={
-            currentTask.terminal !== false &&
-            currentTask.strategy != null &&
-            currentTask.strategy.type === "terminal"
-          }
-        />
-      </ModeBox>
-      <ContentBox
-        className={`diagnostic-content ${
-          mode === "terminal" ? "terminal" : ""
-        }`}
-      >
-        {mode === "terminal" ? (
-          <Term xtermInstance={xtermInstance} />
-        ) : (
-          <DiagnosticsContent
-            mode={mode}
-            diagnostics={diagnostics}
-            onJumpTo={onJumpTo}
-          />
-        )}
-      </ContentBox>
-    </Panel>
-  );
-}
-
-type PanelMode = "diagnostics" | "terminal";
-
-type DiagnosticsModeState = {
-  mode: "diagnostics",
-};
-
-type TerminalModeState = {
-  mode: "terminal",
-};
-
-type PanelState = TerminalModeState | DiagnosticsModeState;
-
-const ModeAdapter = withProps(
-  ({
-    panelState,
-    setPanelState = () => {},
-  }: {
-    panelState: PanelState,
-    setPanelState: (state: PanelState) => void,
-  }) => ({
-    mode: panelState.mode,
-    onModeSelection: (mode: PanelMode): void => {
-      if (mode === "terminal") {
-        setPanelState({
-          mode: "terminal",
-        });
-      } else if (mode === "diagnostics") {
-        setPanelState({
-          mode: "diagnostics",
-        });
-      } else {
-        setPanelState({
-          mode: "diagnostics",
-        });
-      }
-    },
-  }),
-);
-
-export const enhance = compose(
-  withState(
-    "panelState",
-    "setPanelState",
-    ({ currentTask }: { currentTask: Task }): PanelState => {
-      let mode: PanelState;
-
-      if (
-        currentTask.terminal !== false &&
-        currentTask.strategy != null &&
-        currentTask.strategy.type === "terminal"
-      ) {
-        mode = { mode: "terminal" };
-      } else {
-        mode = { mode: "diagnostics" };
-      }
-
-      return mode;
-    },
-  ),
-  ModeAdapter,
-);
-
-export default enhance(DiagnosticsPanel);
